Move MainActivity to main package
