+++ app/src/androidTest/java/com/qwertyfinger/androidsimpleboilerplate/ExampleInstrumentedTest.kt
@@ -32,6 +32,7 @@ import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.qwertyfinger.androidsimpleboilerplate.main.MainActivity
 import org.junit.Test
 import org.junit.runner.RunWith
